+++ NEWS.md
@@ -1,3 +1,9 @@
+# migest 2.0.2
+
+* Added method option to `birth_mat()`
+* Added `sum_bilat()` for `sum_bilateral()`
+* Added `sum_country()`
+
 # migest 2.0.1
 
 * Added a `NEWS.md` file to track changes to the package.
@@ -8,5 +14,4 @@
 * Added `mig_chord()`
 * Added `dict_ims`
 * Replace `sum_turnover()` with `sum_region()`
-* Replace `counter()` with `sum_bilat()`
-
+* Replace `counter()` with `sum_bilateral()`